Factor out some of the common code for thread
