+++ glue/viewers/profile/layer_artist.py
@@ -8,7 +8,7 @@ import numpy as np
 
 from glue.external.six.moves import queue
 from glue.core import Data
-from glue.utils import defer_draw, nanmin, nanmax, queue_to_list
+from glue.utils import defer_draw, nanmin, nanmax
 from glue.viewers.profile.state import ProfileLayerState
 from glue.viewers.matplotlib.layer_artist import MatplotlibLayerArtist
 from glue.core.exceptions import IncompatibleAttribute, IncompatibleDataException
@@ -19,29 +19,7 @@ except Exception:
     QT_INSTALLED = False
 else:
     QT_INSTALLED = True
-
-if QT_INSTALLED:
-
-    # When using Qt, we make use of a thread that continuously listens for
-    # requests to update the profile and we run these as needed. In future,
-    # we should add the ability to interrupt compute jobs if a newer compute
-    # job is requested.
-
-    from qtpy.QtCore import Signal, QThread
-
-    class ComputeWorker(QThread):
-
-        compute_start = Signal()
-        compute_end = Signal()
-        compute_error = Signal(object)
-
-        def __init__(self, function):
-            super(ComputeWorker, self).__init__()
-            self.function = function
-            self.running = False
-
-        def run(self):
-            self.function()
+    from glue.viewers.matplotlib.qt.compute_worker import ComputeWorker
 
 
 class ProfileLayerArtist(MatplotlibLayerArtist):
@@ -80,7 +58,7 @@ class ProfileLayerArtist(MatplotlibLayerArtist):
     def remove(self):
         super(ProfileLayerArtist, self).remove()
         if QT_INSTALLED and self._worker is not None:
-            self._work_queue.put('stop')
+            self._worker.work_queue.put('stop')
             self._worker.exit()
             # Need to wait otherwise the thread will be destroyed while still
             # running, causing a segmentation fault
@@ -97,53 +75,17 @@ class ProfileLayerArtist(MatplotlibLayerArtist):
         self._last_layer_state = {}
 
     def setup_thread(self):
-        self._worker = ComputeWorker(self._thread_loop)
+        self._worker = ComputeWorker(self._calculate_profile_thread)
         self._worker.compute_end.connect(self._calculate_profile_postthread)
         self._worker.compute_error.connect(self._calculate_profile_error)
         self._worker.compute_start.connect(self.notify_start_computation)
-        self._work_queue = queue.Queue()
+        self._worker.work_queue = queue.Queue()
         self._worker.start()
-
-    def _thread_loop(self):
-
-        error = None
-
-        while True:
-
-            time.sleep(1 / 25)
-
-            msgs = queue_to_list(self._work_queue)
-
-            if 'stop' in msgs:
-                return
-            elif len(msgs) == 0:
-                # We change this here rather than in the try...except below
-                # to avoid stopping and starting in quick succession.
-                if self._worker.running:
-                    self._worker.running = False
-                    if error is None:
-                        self._worker.compute_end.emit()
-                    else:
-                        self._worker.compute_error.emit(error)
-                        error = None
-                continue
-
-            # If any resets were requested, honor this
-            reset = any(msgs)
-
-            try:
-                self._worker.running = True
-                self._worker.compute_start.emit()
-                self._calculate_profile_thread(reset=reset)
-            except Exception:
-                error = sys.exc_info()
-            else:
-                error = None
 
     @defer_draw
     def _calculate_profile(self, reset=False):
         if QT_INSTALLED:
-            self._work_queue.put(reset)
+            self._worker.work_queue.put(reset)
         else:
             try:
                 self.notify_start_computation()
